tts_test/pyproject.toml
[project]
name = "tts-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "melotts",
    "python-mecab-ko>=1.3.7",
    "sounddevice>=0.5.2",
]

[tool.uv.sources]
melotts = { path = "MeloTTS" }
